Feature/textures (#21)
* Moved example components into separate lib.
* Create separate example for the sponza model
* add uv vector to mesh class
* Specify and visualize UV coordinates for a cube
* Add Texture class
* Refactor texturing further. Add texture sampling shader.

+++ FIRE/src/VertexDeclaration.cpp
@@ -2,13 +2,13 @@
 
 namespace FIRE
 {
-VertexDeclarationSection::VertexDeclarationSection(size_t size, size_t offset)
+VertexDeclarationSection::VertexDeclarationSection(uint32_t size, uint32_t offset)
     : size(size)
     , offset(offset)
 {
 }
 
-void VertexDeclaration::AddSection(std::string name, size_t size, size_t offset)
+void VertexDeclaration::AddSection(std::string name, uint32_t size, uint32_t offset)
 {
     m_sections.insert(std::make_pair(name, VertexDeclarationSection(size, offset)));
 }
